feat(loadtest): recirculating-contention pilot harness
Ahead of the full recirculating sweep: adds the pieces needed to run a
short hold + sweeper workload and verify it actually recirculates before
committing another ~20 minutes to the full matrix.

app/infra/config.py: hold_duration_seconds widened from int to float --
the benchmark needs sub-integer overrides (e.g. 1.0s), 480.0 (8 minutes)
remains the actual product default.

loadtest/run_benchmark.py: start_api gains an optional
hold_duration_seconds override, passed through as HOLD_DURATION_SECONDS
-- unused by every existing caller (comparison mode, the coarse sweep),
only the new recirculating harness sets it.

app/api/routes/admin.py: GET /api/admin/seat-status-counts (AVAILABLE/
HELD/BOOKED counts for an event) -- real, reusable observability, not a
benchmark-only endpoint, though NOT what the pilot itself polls with (see
next paragraph).

loadtest/recirculating_pilot.py: starts an API instance (short,
overridden hold duration) and workers/sweeper_worker.py (short,
overridden interval) together, seeds a small seat pool, runs
contention_sweep.js, and polls available-seat-count throughout. First
version polled the admin endpoint above over HTTP and got only ~1 sample
every 2s instead of the intended 50ms -- confirmed by direct measurement
that this was the API's OWN connection pool being saturated by the same
200-VU load being observed, not Postgres itself (~60 of 100 default
connections were already in use). Fixed by polling the database directly
through a small dedicated connection pool, bypassing the saturated one
entirely and avoiding adding the poll itself to the measured load.
Reports recirculation cycles observed, fraction of the run with >= 1
seat AVAILABLE, and stdev of available count -- see the pilot results
posted in conversation for the parameters chosen and why.

Runs pessimistic AND optimistic at each ratio, and asserts hold
duration/sweeper interval/batch size are identical across both -- the
sweeper is a third writer against the same table as whichever strategy
is under test, and must never vary by strategy.

+++ app/infra/config.py
@@ -33,7 +33,12 @@ class Settings(BaseSettings):
     naive_race_window_ms: int = 0
 
     # How long a hold lasts once acquired (SPEC.md section 5: 8 minutes).
-    hold_duration_seconds: int = 480
+    # float, not int: the Phase 3 recirculating-contention benchmark
+    # overrides this to sub-integer values (e.g. 1.5s) so inventory
+    # cycles fast enough to observe within a short load-test burst -- a
+    # benchmarking configuration, never a product one. 480 (8 minutes)
+    # remains the actual product default.
+    hold_duration_seconds: float = 480.0
 
     # Uvicorn worker count for `make run-api` (see Makefile). Each worker
     # is a separate process with its own connection pool, so the real
